restructure codes, remove RoleLists and RoleTables
